requirements.txt
pandas==1.3.4
pandas-gbq==0.19.1
prefect==2.7.10
prefect-gcp==0.2.5
google-cloud-storage==2.7.0
beautifulsoup4==4.10.0
python-dotenv==1.0.0